backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
web3==6.15.1
eth-account==0.11.2
python-dotenv==1.0.1
pyjwt==2.8.0
httpx==0.27.0
pydantic==2.6.3
requests==2.31.0
pytest==8.1.1
